+++ HISTORY.md
@@ -1,5 +1,12 @@
 # Version history
 
+## V0.4 2014/06/20
+
+* Added primaryIndent, secondaryIndent and indentCycle options to Antlr4TestTask
+  to correspond to antlr4-regressionTestRig v0.3
+* updated verion (now v0.3) of the antlr4-regressionTestRig dependency in
+  Antlr4TestPlugin.
+
 ## V0.3 2014/06/14
 
 * corrected version (now v0.2) in antlr4-regressionTestRig dependency in